Refactor code and fix bugs

+++ cpp/2022/prelim/p11.cpp
@@ -5,7 +5,7 @@ using namespace std;
 int main() {
     string str;
     cin >> str;
-    int value = stoi(to_string(str[1]) + str[3] + str[2] + str[0]);
+    long long value = stoll(to_string(str[1]) + str[3] + str[2] + str[0]);
     for (int i = 0; i < 2; ++i) {
         value *= value;
         value %= 1'000'000;
